# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ademir API",
    "description": "API usada pelo novo frontend (React/Vite). Endpoints sob o prefixo /api.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3101/api" }
  ],
...
  "paths": {
...
    "/cartoes": {
      "get": { "tags": ["cartoes"], "summary": "Listar cartões", "responses": { "200": { "description": "Lista" } } },
      "post": {
        "tags": ["cartoes"],
        "summary": "Salvar cartões (array completo)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "array" } } } },
        "responses": { "200": { "description": "Cartões salvos" } }
      }
    },
    "/cartoes/{numero}": {
      "get": {
        "tags": ["cartoes"],
        "summary": "Buscar cartão por número",
        "parameters": [{ "in": "path", "name": "numero", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Cartão" }, "404": { "description": "Não encontrado" } }
      }
    },
    "/cartoes/{id}/debitar": {
      "post": {
        "tags": ["cartoes"],
        "summary": "Debitar saldo",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "valor": { "type": "number" }, "itens": { "type": "array" } } } } } },
        "responses": { "200": { "description": "Saldo debitado" }, "400": { "description": "Saldo insuficiente" }, "404": { "description": "Cartão não encontrado" } }
      }
    },
    "/cartoes/{id}/recarregar": {
      "post": {
        "tags": ["cartoes"],
        "summary": "Recarregar saldo",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "valor": { "type": "number" }, "descricao": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Saldo recarregado" }, "404": { "description": "Cartão não encontrado" } }
      }
    },
    "/cartoes/{id}/transacoes": {
      "get": {
        "tags": ["cartoes"],
        "summary": "Listar transações de um cartão",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Lista de transações" } }
      }
    },
    "/cartoes/{id}/estornar": {
      "post": {
        "tags": ["cartoes"],
        "summary": "Estornar transações",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "transacoesIds": { "type": "array", "items": { "type": "integer" } } } } } } },
        "responses": { "200": { "description": "Estorno realizado" }, "400": { "description": "Nenhuma transação válida" }, "404": { "description": "Cartão não encontrado" } }
      }
    },
...
  }
}